+++ py-polars/CHANGELOG.md
@@ -7,12 +7,14 @@ The Rust crate `polars` has its own changelog.
 * name change: Python bindings package renamed from py-polars to polars
 
 * feature
+  - \[python\] eager: DataFrame fold for horizontal aggregation.
   - \[python\] add parquet compression
   - \[python\] shift_and_fill expression
   - \[python\] implicitly download raw files from the web in `read_parquet`, `read_csv`.
   - \[python | rust\] methods for local peak finding in numerical series
   - \[python | rust\] faster query optimization due to local memory arena's.
   - \[rust\] reduce default compile time by making less features default.
+  - \[python | rust\] Series zip_with implicitly cast to supertype.
   
 * bug fix
   - \[python\] support file buffers for reading and writing csv and parquet